+++ Etheria/entities/BulletEntity.vb
@@ -3,7 +3,7 @@
 Public Class BulletEntity
     Inherits Entity
 
-    Public Sub New(spriteName As String, velocity As Vector2f, initialPosition As Vector2f, Optional scale As Single = 0.3)
+    Public Sub New(spriteName As String, velocity As Vector2f, initialPosition As Vector2f, Optional scale As Double = 0.3)
         AddComponents(
             New PositionComponent(initialPosition),
             New VelocityComponent(velocity),
